# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bird",
    "description": "Deploy and run containers on a bird server. Browse these docs at /docs. Every endpoint except the docs needs the API token from birdd's data directory as a bearer token.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/projects/{project}/environments/{environment}/services/{name}/backups": {
      "get": {
        "tags": [
          "backups"
        ],
        "summary": "List backups of a service, newest first",
        "operationId": "list",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Service name, also of a removed service",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Backups, newest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BackupInfo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "backups"
        ],
        "summary": "Back up a service's volumes",
        "operationId": "create",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Service name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Every volume copied while the machines were paused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupInfo"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Service not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The service has no volumes, or another operation is in progress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project}/environments/{environment}/services/{name}/backups/schedule": {
      "put": {
        "tags": [
          "backups"
        ],
        "summary": "Back up a service's volumes on a schedule",
        "operationId": "set_schedule",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Service name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupSchedule"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Schedule saved; the first backup runs within a minute when none was scheduled before",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupSchedule"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Service not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The service has no volumes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "backups"
        ],
        "summary": "Stop backing up a service on a schedule, keeping the backups it made",
        "operationId": "clear_schedule",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Service name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Schedule removed"
          },
          "401": {
            "description": "Missing or invalid API token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Service not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project}/environments/{environment}/services/{name}/backups/{id}": {
      "delete": {
        "tags": [
          "backups"
        ],
        "summary": "Delete a backup and its archives",
        "operationId": "remove",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Service name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Backup id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Backup deleted"
          },
          "401": {
            "description": "Missing or invalid API token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Backup not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project}/environments/{environment}/services/{name}/backups/{id}/restore": {
      "post": {
        "tags": [
          "backups"
        ],
        "summary": "Replace a service's volume data with a backup",
        "operationId": "restore",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "path",
            "description": "Environment name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "path",
            "description": "Service name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Backup id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Data restored and machines started again; the replaced data is kept as a new backup",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestoreResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Service or backup not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "The backup does not fit the service, or another operation is in progress",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BackupId": {
        "type": "string",
        "format": "uuid"
      },
      "BackupInfo": {
        "type": "object",
        "required": [
          "id",
          "service",
          "trigger",
          "storage",
          "volumes",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix seconds"
          },
          "id": {
            "$ref": "#/components/schemas/BackupId"
          },
          "service": {
            "$ref": "#/components/schemas/Name"
          },
          "storage": {
            "type": "string",
            "description": "Where the archives are kept, such as local"
          },
          "trigger": {
            "$ref": "#/components/schemas/BackupTrigger",
            "description": "manual, or restore for the copy taken automatically before a restore"
          },
          "volumes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BackupVolumeInfo"
            }
          }
        }
      },
      "BackupInterval": {
        "type": "integer",
        "format": "int32",
        "description": "Seconds between backups",
        "maximum": 2592000,
        "minimum": 3600
      },
      "BackupKeep": {
        "type": "integer",
        "format": "int32",
        "maximum": 1000,
        "minimum": 1
      },
      "BackupSchedule": {
        "type": "object",
        "required": [
          "every",
          "keep"
        ],
        "properties": {
          "every": {
            "$ref": "#/components/schemas/BackupInterval"
          },
          "keep": {
            "$ref": "#/components/schemas/BackupKeep"
          }
        }
      },
      "BackupTrigger": {
        "type": "string",
        "enum": [
          "manual",
          "restore",
          "scheduled"
        ]
      },
      "BackupVolumeInfo": {
        "type": "object",
        "required": [
          "name",
          "size_bytes"
        ],
        "properties": {
          "name": {
            "$ref": "#/components/schemas/Name"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "logs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Name": {
        "type": "string"
      },
...
      "RestoreRequest": {
        "type": "object",
        "properties": {
          "allow_image_change": {
            "type": "boolean",
            "description": "Restore data written by a different image than the service runs now"
          }
        }
      },
      "RestoreResponse": {
        "type": "object",
        "required": [
          "restored",
          "safety_backup"
        ],
        "properties": {
          "restored": {
            "$ref": "#/components/schemas/BackupId"
          },
          "safety_backup": {
            "$ref": "#/components/schemas/BackupId",
            "description": "Backup of the data that was replaced, to undo the restore"
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_token": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "api_token": []
    }
  ],
...
}